requirements.txt
python-pillow
requests
pillow
pytest
